[issue13] Revert "Merge branch 'fresh-opensrp-profiling-with-refactor' into OpenSrpSimplifyRegistry"
This reverts commit 249ea820bc6ce199a8bbee314c681bffd99d9d67, reversing
changes made to 70b25cce824ab03672ac05bfcd112e2f360dce45.
