--- dristhi-app/assets/app-extra.properties
@@ -1,1 +1,0 @@
-AVAILABLE_REGISTERS=WAIT_FOR_VALUE
