+++ dristhi-app/assets/template/child_detail.handlebars
