+++ dristhi-app/assets/www/smart_registry/child_register.html
@@ -1,27 +1,27 @@
 <!doctype html>
 <html ng-app="smartRegistry">
 <head>
-    <link href="../../../../../assets/css/smart_registry/reset.css" media="all" rel="stylesheet" type="text/css"/>
-    <link href="../../../../../assets/css/smart_registry/styles.css" media="all" rel="stylesheet" type="text/css"/>
-    <link href="../../../../../assets/css/smart_registry/smart_registry.css" media="all" rel="stylesheet" type="text/css"/>
-    <link href="../../../../../assets/css/flat-bootstrap.css" rel="stylesheet">
-    <link href="../../../../../assets/css/custom.css" rel="stylesheet">
-    <script src="../../../../../assets/js/lib/angular.min.js" type="text/javascript"></script>
-    <script src="../../../../../assets/js/lib/ui-bootstrap-tpls-0.3.0.min.js"></script>
-    <script src="../../../../../assets/js/page_view.js"></script>
-    <script src="../../../../../assets/js/village.js"></script>
-    <script src="../../../../../assets/js/anmNavigation.js" type="text/javascript"></script>
-    <script src="../../../../../assets/js/form_context.js" type="text/javascript"></script>
-    <script src="../../../../../assets/js/smart_registry/app.js" type="text/javascript"></script>
-    <script src="../../../../../assets/js/smart_registry/smart_filters.js" type="text/javascript"></script>
-    <script src="../../../../../assets/js/smart_registry/smart_helper_service.js" type="text/javascript"></script>
-    <script src="../../../../../assets/js/smart_registry/smart_directives.js" type="text/javascript"></script>
-    <script src="../../../../../assets/js/smart_registry/debounce_service.js" type="text/javascript"></script>
-    <script src="../../../../../assets/js/smart_registry/child_service.js" type="text/javascript"></script>
-    <script src="../../../../../assets/js/smart_registry/list_view_controller.js" type="text/javascript"></script>
-    <script src="../../../../../assets/js/smart_registry/child_registry_bridge.js" type="text/javascript"></script>
-    <script src="../../../../../assets/js/smart_registry/anm_location_bridge.js" type="text/javascript"></script>
-    <script src="../../../../../assets/js/smart_registry/child_register_app.js" type="text/javascript"></script>
+    <link href="../../css/smart_registry/reset.css" media="all" rel="stylesheet" type="text/css"/>
+    <link href="../../css/smart_registry/styles.css" media="all" rel="stylesheet" type="text/css"/>
+    <link href="../../css/smart_registry/smart_registry.css" media="all" rel="stylesheet" type="text/css"/>
+    <link href="../../css/flat-bootstrap.css" rel="stylesheet">
+    <link href="../../css/custom.css" rel="stylesheet">
+    <script src="../../js/lib/angular.min.js" type="text/javascript"></script>
+    <script src="../../js/lib/ui-bootstrap-tpls-0.3.0.min.js"></script>
+    <script src="../../js/page_view.js"></script>
+    <script src="../../js/village.js"></script>
+    <script src="../../js/anmNavigation.js" type="text/javascript"></script>
+    <script src="../../js/form_context.js" type="text/javascript"></script>
+    <script src="../../js/smart_registry/app.js" type="text/javascript"></script>
+    <script src="../../js/smart_registry/smart_filters.js" type="text/javascript"></script>
+    <script src="../../js/smart_registry/smart_helper_service.js" type="text/javascript"></script>
+    <script src="../../js/smart_registry/smart_directives.js" type="text/javascript"></script>
+    <script src="../../js/smart_registry/debounce_service.js" type="text/javascript"></script>
+    <script src="../../js/smart_registry/child_service.js" type="text/javascript"></script>
+    <script src="../../js/smart_registry/list_view_controller.js" type="text/javascript"></script>
+    <script src="../../js/smart_registry/child_registry_bridge.js" type="text/javascript"></script>
+    <script src="../../js/smart_registry/anm_location_bridge.js" type="text/javascript"></script>
+    <script src="../../js/smart_registry/child_register_app.js" type="text/javascript"></script>
 </head>
 <body ng-controller="childRegisterController">
 <div ng-controller="listViewController">
