--- dristhi-app/registers/child/assets/www/child_detail.html
@@ -1,48 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-<head>
-    <link href="../../../../assets/css/bootstrap.css" rel="stylesheet">
-    <link href="../../../../assets/css/custom.css" rel="stylesheet">
-    <link href="../../../../assets/css/timeline.css" rel="stylesheet">
-
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=1.0, user-scalable=no">
-</head>
-
-<body>
-
-<div class="page">
-    <div id="mainpanel" class="mainpanel"></div>
-</div>
-
-</body>
-
-<script src="../../../../assets/js/lib/jquery.min.js"></script>
-<script src="../../../../assets/js/lib/jquery-ui-1.8.23.custom.min.js"></script>
-<script src="../../../../assets/js/lib/bootstrap.min.js"></script>
-<script src="../../../../assets/js/lib/handlebars.runtime.js"></script>
-
-<script src="../../../../assets/js/custom.ui.js"></script>
-<script src="../../../../assets/template/templates.js"></script>
-<script src="../../../../assets/js/modal.js"></script>
-<script src="../../../../assets/js/international_text.js"></script>
-<script src="../../../../assets/js/page_view.js"></script>
-<script src="../../../../assets/js/form_context.js"></script>
-
-<script src="../../../../assets/js/child_detail.js"></script>
-<script>
-    $(document).ready(function () {
-        var child = new Child(new ChildBridge(), new FormBridge());
-        child.populateInto("#mainpanel");
-        child.bindEveryItemToForm(".modal-row");
-        child.bindToCamera("#childPhoto");
-        child.bindPhoto("#childPhoto");
-
-        pageView.onReloadPhoto(function (caseId, photoPath) {
-            child.reloadPhoto("#childPhoto", caseId, photoPath);
-        });
-
-        new Modal().bindToClick();
-    });
-</script>
-
-</html>
