+++ dristhi-app/res/layout/child_register_client_sick_status.xml
