+++ dristhi-app/res/layout/child_smart_register_id_details.xml
