+++ dristhi-app/res/layout/smart_registers_home.xml
@@ -38,7 +38,6 @@
                 android:orientation="horizontal">
 
             <FrameLayout
-                    android:id="@+id/enter_ec_register"
                     android:layout_width="0dp"
                     android:layout_height="fill_parent"
                     android:layout_weight="1">
@@ -71,7 +70,6 @@
             </FrameLayout>
 
             <FrameLayout
-                    android:id="@+id/enter_fp_register"
                     android:layout_width="0dp"
                     android:layout_height="fill_parent"
                     android:layout_weight="1"
@@ -104,7 +102,6 @@
             </FrameLayout>
 
             <FrameLayout
-                    android:id="@+id/enter_anc_register"
                     android:layout_width="0dp"
                     android:layout_height="fill_parent"
                     android:layout_weight="1">
@@ -136,7 +133,6 @@
             </FrameLayout>
 
             <FrameLayout
-                    android:id="@+id/enter_pnc_register"
                     android:layout_width="0dp"
                     android:layout_height="fill_parent"
                     android:layout_weight="1">
@@ -168,7 +164,6 @@
             </FrameLayout>
 
             <FrameLayout
-                    android:id="@+id/enter_child_register"
                     android:layout_width="0dp"
                     android:layout_height="fill_parent"
                     android:layout_weight="1">
