+++ dristhi-app/src/main/java/org/ei/drishti/provider/ChildSmartRegisterClientsProvider.java
