+++ dristhi-app/src/main/java/org/ei/drishti/view/activity/ChildSmartRegisterActivity.java
