+++ dristhi-app/src/main/java/org/ei/drishti/view/contract/ChildSmartRegisterClient.java
