+++ dristhi-app/src/main/java/org/ei/drishti/view/controller/ChildSmartRegisterController.java
@@ -20,8 +20,6 @@ import java.util.List;
 import static java.lang.String.valueOf;
 import static java.util.Collections.sort;
 import static org.apache.commons.lang3.StringUtils.isBlank;
-import static org.ei.drishti.AllConstantsForChildRegister.ChildRegistrationFields.NAME;
-import static org.ei.drishti.AllConstantsForChildRegister.ChildRegistrationFields.WEIGHT;
 import static org.ei.drishti.domain.ServiceProvided.*;
 
 public class ChildSmartRegisterController {
@@ -56,9 +54,9 @@ public class ChildSmartRegisterController {
                             new ChildClient(
                                     child.caseId(),
                                     child.gender(),
-                                    child.getDetail(WEIGHT),
+                                    child.getDetail(AllConstants.ChildRegistrationFields.WEIGHT),
                                     child.mother().thayiCardNumber())
-                                    .withName(child.getDetail(NAME))
+                                    .withName(child.getDetail(AllConstants.ChildRegistrationFields.NAME))
                                     .withEntityIdToSavePhoto(child.caseId())
                                     .withMotherName(child.ec().wifeName())
                                     .withDOB(child.dateOfBirth())
@@ -97,9 +95,9 @@ public class ChildSmartRegisterController {
                             new ChildClient(
                                     child.caseId(),
                                     child.gender(),
-                                    child.getDetail(WEIGHT),
+                                    child.getDetail(AllConstants.ChildRegistrationFields.WEIGHT),
                                     child.mother().thayiCardNumber())
-                                    .withName(child.getDetail(NAME))
+                                    .withName(child.getDetail(AllConstants.ChildRegistrationFields.NAME))
                                     .withEntityIdToSavePhoto(child.caseId())
                                     .withMotherName(child.ec().wifeName())
                                     .withDOB(child.dateOfBirth())
