+++ dristhi-app/src/main/java/org/ei/drishti/view/viewHolder/ChildRegisterProfilePhotoLoader.java
